Intro: Explain why we turned WiFi off
