+++ packages/thicket-intro/src/App/GifCreator/Image.js
@@ -1,0 +1,7 @@
+import React from 'react'
+import styled from 'styled-components'
+
+export default styled.img`
+  height: 2.5em;
+`
+
